// src/app/dashboard/interfaces/books-response.ts
 export interface Book {
  id?: string;
  title?: string;
  author?: string;
  publicationDate?: string;
  gender?: string;
  publisher?: string;
  shelf?: string;
}



export interface BooksResponse {
  content: Book[];
  pageable: {
    pageNumber: number;
    pageSize: number;
    sort: {
      empty: boolean;
      sorted: boolean;
      unsorted: boolean;
    };
    offset: number;
    unpaged: boolean;
    paged: boolean;
  };
  last: boolean;
  totalPages: number;
  totalElements: number;
  first: boolean;
  size: number;
  number: number;
  sort: {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };
  numberOfElements: number;
  empty: boolean;
}